+++ bases/main/test/psi/launcher_test.clj
@@ -1,5 +1,6 @@
 (ns psi.launcher-test
   (:require
+   [clojure.java.io]
    [clojure.test :refer [deftest is testing]]
    [psi.launcher :as launcher]
    [psi.launcher.extensions :as extensions]))
@@ -64,19 +65,24 @@
                nrepl/nrepl {:mvn/version "1.5.1"}}
              (:deps (with-redefs [launcher/repo-basis-config (constantly repo-config)]
                       (launcher/psi-self-basis "/repo/psi" :installed))))))
-    (testing "jar policy emits single mvn coord for the release version"
+    (testing "jar policy reads external runtime deps from jar-owned release metadata and retains the psi artifact coord"
       (is (= '{org.hugoduncan/psi {:mvn/version "0.1.42"}
-               nrepl/nrepl {:mvn/version "1.5.1"}}
-             (:deps (with-redefs [launcher/release-version (constantly "0.1.42")]
+               nrepl/nrepl {:mvn/version "1.5.1"}
+               io.github.timokramer/charm.clj {:git/tag "v0.2.71"
+                                               :git/sha "38e6823"}}
+             (:deps (with-redefs [launcher/release-version (constantly "0.1.42")
+                                  launcher/release-basis-config (constantly '{:deps {nrepl/nrepl {:mvn/version "1.5.1"}
+                                                                                     io.github.timokramer/charm.clj {:git/tag "v0.2.71"
+                                                                                                                     :git/sha "38e6823"}}})]
                       (launcher/psi-self-basis "/repo/psi" :jar))))))
-    (testing "jar policy throws when version resource is unreleased"
-      (let [ex (try
-                 (with-redefs [launcher/release-version (constantly nil)]
+    (testing "jar policy accepts the baked version string, including local unreleased smoke installs"
+      (is (= '{org.hugoduncan/psi {:mvn/version "unreleased"}
+               nrepl/nrepl {:mvn/version "1.5.1"}}
+             (-> (with-redefs [launcher/release-version (constantly "unreleased")
+                               launcher/release-basis-config (constantly '{:deps {nrepl/nrepl {:mvn/version "1.5.1"}}})]
                    (launcher/psi-self-basis "/repo/psi" :jar))
-                 nil
-                 (catch clojure.lang.ExceptionInfo e e))]
-        (is (some? ex))
-        (is (= :basis-construction (-> ex ex-data :stage)))))))
+                 :deps
+                 (select-keys '[org.hugoduncan/psi nrepl/nrepl])))))))
 
 (deftest build-deps-clj-args-test
   (let [args (launcher/build-deps-clj-args {:basis-file "/tmp/psi-basis-123.edn"
@@ -122,11 +128,9 @@
                          :defaulted-libs ['psi/mementum 'psi/workflow-loader]
                          :inferred-init-libs ['psi/mementum 'psi/workflow-loader]}
           result (with-redefs [launcher/release-version (constantly "0.1.42")
+                               launcher/release-basis-config (constantly '{:deps {nrepl/nrepl {:mvn/version "1.5.1"}}})
                                launcher/manifest-state (fn [_ _ _] manifest-info)]
                    (launcher/startup-basis "/repo/psi" "/repo/project" :jar))]
-      ;; psi self-dep is the single mvn coord
-      (is (= {:mvn/version "0.1.42"}
-             (get-in result [:basis :deps 'org.hugoduncan/psi])))
       ;; psi-owned extensions are NOT added as separate Maven artifacts
       (is (nil? (get-in result [:basis :deps 'psi/mementum])))
       (is (nil? (get-in result [:basis :deps 'psi/workflow-loader])))))
@@ -142,29 +146,15 @@
                          :defaulted-libs []
                          :inferred-init-libs []}
           result (with-redefs [launcher/release-version (constantly "0.1.42")
+                               launcher/release-basis-config (constantly '{:deps {nrepl/nrepl {:mvn/version "1.5.1"}}})
                                launcher/manifest-state (fn [_ _ _] manifest-info)]
                    (launcher/startup-basis "/repo/psi" "/repo/project" :jar))]
       (is (= {:mvn/version "2.0.0"}
-             (get-in result [:basis :deps 'third-party/ext])))))
-  (testing "jar policy throws when version resource is unreleased"
-    (let [manifest-info {:user-path "/tmp/user.edn"
-                         :project-path "/tmp/project/.psi/extensions.edn"
-                         :user-present? false
-                         :project-present? false
-                         :user-manifest {:deps {}}
-                         :project-manifest {:deps {}}
-                         :merged-manifest {:deps {}}
-                         :expanded-manifest {:deps {}}
-                         :defaulted-libs []
-                         :inferred-init-libs []}
-          ex (try
-               (with-redefs [launcher/release-version (constantly nil)
-                             launcher/manifest-state (fn [_ _ _] manifest-info)]
-                 (launcher/startup-basis "/repo/psi" "/repo/project" :jar))
-               nil
-               (catch clojure.lang.ExceptionInfo e e))]
-      (is (some? ex))
-      (is (= :basis-construction (-> ex ex-data :stage))))))
+             (get-in result [:basis :deps 'third-party/ext])))
+      (is (= {:mvn/version "1.5.1"}
+             (get-in result [:basis :deps 'nrepl/nrepl])))
+      (is (= {:mvn/version "0.1.42"}
+             (get-in result [:basis :deps 'org.hugoduncan/psi]))))))
 
 (deftest startup-basis-expands-recognized-psi-owned-minimal-manifest-entry-into-basis-deps
   (let [repo-config {:deps {'psi/main {:local/root "bases/main"}
@@ -188,6 +178,17 @@
            (get-in result [:basis :deps 'psi/workflow-loader])))
     ;; :psi/init is also stripped from manifest-info expanded-manifest (same expanded-deps map)
     (is (nil? (get-in result [:manifest-info :expanded-manifest :deps 'psi/workflow-loader :psi/init])))))
+
+(deftest release-basis-config-test
+  (testing "missing jar-owned release metadata fails clearly"
+    (let [ex (try
+               (with-redefs [clojure.java.io/resource (constantly nil)]
+                 (launcher/release-basis-config))
+               nil
+               (catch clojure.lang.ExceptionInfo e e))]
+      (is (some? ex))
+      (is (= :basis-construction (-> ex ex-data :stage)))
+      (is (= :jar (-> ex ex-data :policy))))))
 
 (deftest launch-plan-test
   (let [basis-state {:basis {:deps {'foo/bar {:mvn/version "1.0.0"}}}